bot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==21.7
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pydantic==2.9.2
pydantic-settings==2.1.0
semantic-kernel==1.37.0
httpx==0.27.2
python-dotenv==1.0.0
ollama==0.4.4
alembic==1.13.1
aio-pika==9.4.3
cryptography>=41.0.0